requirements.txt
# Core Streamlit framework
streamlit>=1.20.0

# Data validation
pydantic>=2.0.0

# Charts and visualization
plotly>=5.0.0

# Data handling
pandas>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Database (built-in sqlite3, but adding for backup)
# sqlite3 is included with Python

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0